made changes to the encryption and decryption algorithm to check if file exists

+++ Cryptoboy.py
@@ -1,9 +1,11 @@
 import string
+import os
 
 
 def encrypt(file, key, outfile):
     """
-    Encrypts the contents of an input file using the Caesar cipher encryption algorithm and writes the ciphertext to an output file.
+    Encrypts the contents of an input file using the Caesar cipher encryption algorithm and writes the ciphertext to
+    an output file.
 
     Args:
     input_file (str): The name of the input file to encrypt.
@@ -13,29 +15,35 @@ def encrypt(file, key, outfile):
 
     alphabet = list(string.ascii_uppercase)
 
-    with open(file, 'r') as f_in:
-        plaintext = f_in.read()
+    # checks to see if file exists within the current directory if it does it continue with the encryption,
+    # if not it prints out the error message
+    if not os.path.exists(file):
+        print(f'{file} does not exist in current directory')
 
-    uppertext = ""
-    ciphertext = ''
-    for char in plaintext:
-        uppertext += char.upper()
+    else:
+        with open(file, 'r') as f_in:
+            plaintext = f_in.read()
 
-    for char in uppertext:
-        if char in alphabet:
-            index = alphabet.index(char)
-            index = (index + key) % 26
-            ciphertext += alphabet[index]
-        else:
-            ciphertext += char
+        uppertext = ""
+        ciphertext = ''
+        for char in plaintext:
+            uppertext += char.upper()
 
-    with open(outfile, 'w') as f_out:
-        f_out.write(ciphertext)
+        for char in uppertext:
+            if char in alphabet:
+                index = alphabet.index(char)
+                index = (index + key) % 26
+                ciphertext += alphabet[index]
+            else:
+                ciphertext += char
+
+        with open(outfile, 'w') as f_out:
+            f_out.write(ciphertext)
 
 
 def decrypt(file, key, outfile):
     """
-    Decrypts the contents of an input file using the Caesar cipher encryption algorithm and writes the ciphertext to an output file.
+    Decrypts the contents of an input file using the Decryption algorithm and outputs to the specified file name.
 
     Args:
     input_file (str): The name of the input file to encrypt.
@@ -45,24 +53,30 @@ def decrypt(file, key, outfile):
 
     alphabet = list(string.ascii_uppercase)
 
-    with open(file, 'r') as f_in:
-        plaintext = f_in.read()
+    # checks to see if file exists within the current directory if it does it continue with the encryption,
+    # if not it prints out the error message
+    if not os.path.exists(file):
+        print(f'{file} does not exist in current directory')
 
-    uppertext = ""
-    ciphertext = ''
-    for char in plaintext:
-        uppertext += char.upper()
+    else:
+        with open(file, 'r') as f_in:
+            plaintext = f_in.read()
 
-    for char in uppertext:
-        if char in alphabet:
-            index = alphabet.index(char)
-            index = (index - key) % 26
-            ciphertext += alphabet[index]
-        else:
-            ciphertext += char
+        uppertext = ""
+        ciphertext = ''
+        for char in plaintext:
+            uppertext += char.upper()
 
-    with open(outfile, 'w') as f_out:
-        f_out.write(ciphertext)
+        for char in uppertext:
+            if char in alphabet:
+                index = alphabet.index(char)
+                index = (index - key) % 26
+                ciphertext += alphabet[index]
+            else:
+                ciphertext += char
+
+        with open(outfile, 'w') as f_out:
+            f_out.write(ciphertext)
 
 
 def crack(file, threshold):
